Improve memory Ordering
## This Commit

Downgrades the memory ordering for the `odd_next` bool operations from
`SeqCst` to what is appropriate.

## Why?

For improved performance and clarity. This should continue working after
we remove the `Mutex` around the stream (once we figure out how). This
could very well be too relaxed, but it'd be hard to tell without Loom or
Miri.